+++ clients/YandexOpenId.php
@@ -15,7 +15,7 @@ use yii\authclient\OpenId;
  *
  * Example application configuration:
  *
- * ~~~
+ * ```php
  * 'components' => [
  *     'authClientCollection' => [
  *         'class' => 'yii\authclient\Collection',
@@ -27,7 +27,7 @@ use yii\authclient\OpenId;
  *     ]
  *     ...
  * ]
- * ~~~
+ * ```
  *
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 2.0