+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 authclient extension Change Log
 2.0.6 under development
 -----------------------
 
-
+- Enh #58: Added support for requesting email from Twitter (umanamente)
 
 2.0.5 September 23, 2015
 ------------------------