+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Yii Framework 2 authclient extension Change Log
 - Bug #37: Fixed `\yii\authclient\widgets\AuthChoice` overrides any `<a>` tag click behavior between `begin()` and `end()` methods (klimov-paul)
 - Enh #58: Added support for requesting email from Twitter (umanamente)
 
+
 2.0.5 September 23, 2015
 ------------------------